Develop

+++ packages/features/team-accounts/src/server/actions/services/delete/delete-service-server.tsx
@@ -1,23 +1,26 @@
 'use server';
 
-import { revalidatePath } from 'next/cache';
-
-
-
 import { getSupabaseServerComponentClient } from '@kit/supabase/server-component-client';
+import { getStripeAccountID } from '../../members/get/get-member-account';
 
-export const deleteService = async (serviceId: number) => {
+// const baseUrl = process.env.NEXT_PUBLIC_SITE_URL ?? 'http://localhost:3000';
+
+export const deleteService = async (priceId: string) => {
   try {
+
+    const stripe_account_id = await getStripeAccountID();
+    if (!stripe_account_id) throw new Error('No stripe account found');      
+
     const client = getSupabaseServerComponentClient();
+    // Delete From DB
     const { error } = await client
       .from('services')
       .delete()
-      .eq('id', serviceId);
+      .eq('price_id', priceId);
 
     if (error) {
       throw new Error(error.message);
     }
-    revalidatePath('/services');
   } catch (error) {
     console.error('Error al eliminar el usuario:', error);
     throw error;
